Add html, styles and script to the project
Add html template that contains only a title, a subtitle and a div that
will contain all the 16x16 squares to form a grid.

Add reset.css that overrides default styles from some elements.

Add styles.css for giving styles to the board and its squares.

Add a script that construct a perfect board that is made of 16x16
squares.